+++ CHANGELOG.md
@@ -6,12 +6,14 @@ This is a summary of new features and bugfixes. Read the README to learn how to 
 ## 3.3.0
 
 Jibb added basic support for mapping gyro and flick stick to virtual controller sticks.
+Nicolas is adding Nielk1's trigger effect to JSM.
 
 ### Features
 * GYRO_OUTPUT can be set to RIGHT_STICK to convert gyro to stick instead of mouse. Find it in the README to see other settings available, including "UNDEADZONE", "UNPOWER", and "VIRTUAL_SCALE".
 * FLICK_STICK_OUTPUT can also be set to RIGHT_STICK to fake flick stick in a game that's only reading from the controller. It will usually be less precise than mouse-flick-stick.
 * Tune virtual flick stick and gyro aiming with VIRTUAL_STICK_CALIBRATION instead of REAL_WORLD_CALIBRATION and IN_GAME_SENS.
 * ZL and ZR bindings now work with alongside virtual stick bindings (setting ZL_MODE or ZR_MODE to a virtual conroller trigger).
+* New settings LEFT_TRIGGER_EFFECT and RIGHT_TRIGGER_EFFECT can be set to BOW, SEMI_AUTOMATIC or MACHINE. Type HELP for details
 
 ### Bugfixes
 * Fixed virtual DS4 left stick having a heavy bias to the corners.